backend/scripts/requirements-seed.txt
supabase>=2.10.0
httpx>=0.27.0
python-dotenv>=1.0.0
